@@ -3,8 +3,8 @@ cmake_minimum_required(VERSION 2.8.5)
 project(ntest)
 
 if (CMAKE_COMPILER_IS_GNUCC OR "${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang" OR "${CMAKE_CXX_COMPILER_ID}" STREQUAL "Intel")
-    add_definitions(-pthread -D_M_AMD64 -m64 -msse4 -std=c++0x -O3 -g3 -Wno-unused-result -Wno-deprecated)
-    set(CMAKE_CXX_LINK_FLAGS "-pthread -m64 -msse4 -std=c++0x -O3 -g3")
+    add_definitions(-pthread -D_M_AMD64 -m64 -msse4 -std=c++1y -O3 -g3 -Wno-unused-result -Wno-deprecated)
+    set(CMAKE_CXX_LINK_FLAGS "-pthread -m64 -msse4 -std=c++1y -O3 -g3")
     if("${NTEST_MINGW_BUILD}")
         add_definitions(-static-libgcc -static-libstdc++)
         set(CMAKE_CXX_LINK_FLAGS "${CMAKE_CXX_LINK_FLAGS} -static-libgcc -static-libstdc++")
@@ -21,6 +21,8 @@ if(MSVC)
 	set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} /Ox /GL /MT")
 	set(CMAKE_CXX_FLAGS_RELWITHDEBINFO "${CMAKE_CXX_FLAGS_RELWITHDEBINFO} /Ox /GL /MTd")
 endif(MSVC)
+find_library(TCMALLOC_LIBRARY NAMES tcmalloc_and_profiler tcmalloc
+             HINTS /opt/gperftools/lib)
 
 enable_testing()
 
@@ -33,8 +35,8 @@ add_subdirectory(game)
 file(GLOB HEADER_FILES *.h *.hpp)
 add_library(mainlib GameX.cpp Evaluator.cpp EvalTest.cpp MPCCalc.cpp NtestStream.cpp options.cpp PlayerComputer.cpp Pos2.cpp Pos2Test.cpp Search.cpp SearchTest.cpp SearchParams.cpp SmartBook.cpp SpeedTest.cpp treedebug.cpp Stable.cpp ${HEADER_FILES})
 
-add_executable(ntest main.cpp)
-target_link_libraries(ntest mainlib core game patterns odk n64)
+add_executable(ntest ntest.cpp)
+target_link_libraries(ntest mainlib core game patterns odk n64 ${TCMALLOC_LIBRARY})
 if(MSVC OR "${NTEST_MINGW_BUILD}")
     target_link_libraries(ntest wsock32)
 endif()
