@@ -1,5 +1,5 @@
 #!/usr/bin/python
-# Copyright (c) 2014, Vlad Petric
+# Copyright (c) 2014,2015,2016 Vlad Petric
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
